upload 1

+++ database/migrations/2024_11_06_211925_create_wallets_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
-            $table->integer('balance')->default(0);
+            $table->double('balance')->default(0);
             $table->foreignId('user_id')->constrained();
             $table->timestamps();
         });
